Change lfp_sendfile() return type from int to ssize_t

+++ src/lib/sendfile.c
@@ -35,17 +35,17 @@
 #include <libfixposix.h>
 #include "utils.h"
 
-int lfp_sendfile(int out_fd, int in_fd, off_t offset, size_t nbytes)
+ssize_t lfp_sendfile(int out_fd, int in_fd, off_t offset, size_t nbytes)
 {
 #if defined(HAVE_SENDFILE)
 # if defined(__linux__)
     off_t off = offset;
-    return sendfile(out_fd, in_fd, &off, nbytes);
+    return (ssize_t) sendfile(out_fd, in_fd, &off, nbytes);
 # elif defined(__FreeBSD__)
-    return sendfile(in_fd, out_fd, offset, nbytes, NULL, SF_MNOWAIT);
+    return (ssize_t) sendfile(in_fd, out_fd, offset, nbytes, NULL, SF_MNOWAIT);
 # elif defined(__DARWIN__)
     off_t len = nbytes;
-    return sendfile(in_fd, out_fd, offset, &len, NULL, 0);
+    return (ssize_t) sendfile(in_fd, out_fd, offset, &len, NULL, 0);
 # else
     return ENOSYS;
 # endif
